When fetching relationships, support references that occur inside oneOf subschemas.
If a property schema has oneOf at the top level, then prior to this update, the property was not recognized as a relationship for purposes of expanding references. Now, when building the concrete schema that will be used for reference expansion, we check each oneOf option. If any option describes a relationship with an entity type or storage method (like 'ref' or 'inverse-ref'), we copy those properties to the top level.