Update DAO to expand nested reference paths (#3)
Enables expanding of nested reference paths by adding recursion to `fetchRelationships`.

To accomplish this, the list of reference paths is split into nested and non-nested, with only the non-nested reference paths expanded in the current iteration. The nested reference paths are grouped by prefix, and recursive call is made for each group of reference paths, along with the corresponding entity type and prefix. The prefix is then passed when fetching forward references to calculate the correct path in recursive iterations.