Use fixed order when importing directory of files
When importing modules from a directory, sort files by name to ensure they are loaded in the same order at each start. The order should not matter here, but have a non-deterministic order is resulting in some nested relationships not expanding in specific entity types' calculated "concrete schemas".

The reverse order is being applied to address the first known case where this is happening (relationship `sci-nuclei-aliquot -> sci-powdered-aliquot -> sci-sample`), but is not a reliable fix for all possible cases. A more robust fix to this issue will be to make sure the entity types' loading order does not impact their calculated schemas.