Support circular schemas.
Circular schema references, such as a -> b -> a, were already largely supported, but reference expansion would stop prematurely. This was because an empty schema was cached. (In our example, the cached schema "a" would be empty.) Here we correct this by ensuring that the full parent schema is cached.

With this change, reference expansion in schemas with circularity may go on indefinitely. Additional changes are needed to prevent this.