Allow circular schema references with different storage methods, entity types, or foreign keys.
When expanding a schema, we cache each referenced subschema by name in order to support circular references. The cache has been using the subschema name as a key, which caused problems because the storage method, entity type, and foreign key are merged into the subschema itself, so if the same subschema is referenced twice with different storage methods (for instance), both occurrences in the fully expanded concrete schema will have the same storage method.

To handle this, we add the entity type name, foreign key, and storage method to the cache key.